+++ crates/integrations/connector-integration/src/default_implementations.rs
@@ -885,6 +885,7 @@ default_impl_payment_method_eligibility_v2!(
         Truelayer,
         Hyperswitch,
         Affirm,
+        Maya
     ],
 );